+++ test/boutros/matsu/compiler_test.clj
@@ -1,6 +1,7 @@
 (ns boutros.matsu.compiler-test
   (:require [boutros.matsu.compiler :refer [encode]]
-            [clojure.test :refer :all])
+            [clojure.test :refer :all]
+            [clj-time.core :refer [date-time]])
   (:import (java.net URI)))
 
 (deftest fn-encode
@@ -19,4 +20,6 @@
        ["42" "somedatatype"]        "\"42\"^^somedatatype"
        [:keyword]                  "<keyword>"
        [[]]                        "[]"
-       (range 2)                   '(0 \space 1)))
+       (range 2)                   '(0 \space 1)
+       (date-time 1981 9 8)        "\"1981-09-08T00:00:00.000Z\"^^xsd:dateTime"
+       (java.util.Date. 1)         "\"1970-01-01T00:00:00.001Z\"^^xsd:dateTime"))