+++ CHANGES.md
@@ -1,4 +1,6 @@
 ### 0.1.1-SNAPSHOT
+* feature: compiler encodes from java.util.Date and org.joda.time.DateTime
+           formats to xsd:dateTime
 * feature: pretty-printer
 * feature: compiler encodes lazy-seqs
 * fix: ensure proper escaping of double quotes in strings